requirements.txt
pyspark==3.5.5
delta-spark
mlflow
pandas
